direct comparison to baseline

+++ lib/classifiers/ContextAwareClassifier.py
@@ -138,7 +138,7 @@ class ContextAwareClassifier():
         else:
             n_pos = len([l for l in tr_labs if l == 1])
             class_weight = 1 - (n_pos / len(tr_labs))
-            self.criterion = nn.BCELoss(weight=torch.tensor(class_weight, dtype=torch.float, device=self.device))
+            self.criterion = nn.BCELoss(weight=torch.tensor(class_weight, dtype=torch.long, device=self.device))
 
     def load_model(self, name):
         cpfp = os.path.join(self.cp_dir, name)
